Fix #120 (partially)
This resolves an issue where NbtNumberRanges weren't included correctly in the ingredient previews.
This meant that these preview stacks weren't matching the ingredient.
Based on that the dynamic ingredient resolution failed in these cases and logged errors.

This commit fixes this for NbtNumberRanges.
Since this might reoccur in the future but isn't really a full-on error, the log messages have also been downgraded to warnings.